agent-project/requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
langgraph
langgraph-prebuilt
chromadb
python-dotenv
tavily
sentence-transformers